+++ vck/src/jvmTest/kotlin/TestJavaApiTest.kt
@@ -4,4 +4,7 @@ val TestJavaApiTest by matrixSuite {
     "creates StatusListInfo through the Java API" {
         TestJavaApi().createsStatusListInfoFromJavaApi()
     }
+    "creates nested claims through the Java API" {
+        TestJavaApi().createsNestedClaimFromJavaApi()
+    }
 }